requirements.txt
fastapi[standard]
openai
requests
python-dotenv
streamlit
numpy